feat: Split `main.js` into `cli` and `index`. Index contains importable code, cli contains commandline interface.
